+++ src/Products/PloneMeeting/MeetingItem.py
@@ -6419,7 +6419,6 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         meeting = self._otherMCMeetingToBePresentedIn(destMeetingConfig)
         if meeting:
             newItem.setPreferredMeeting(meeting.UID())
-
         # handle 'otherMeetingConfigsClonableToPrivacy' of original item
         if destMeetingConfigId in self.getOtherMeetingConfigsClonableToPrivacy() and \
            'privacy' in destMeetingConfig.getUsedItemAttributes():
@@ -6465,7 +6464,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                     try:
                         # special handling for the 'present' transition
                         # that needs a meeting as 'PUBLISHED' object to work
-                        if tr == 'present':
+                        if tr == 'present' and \
+                           not isinstance(newItem.wfConditions()._check_required_data("presented"), No):
                             if not meeting:
                                 plone_utils.addPortalMessage(
                                     _('could_not_present_item_no_meeting_accepting_items',