+++ CHANGES.rst
@@ -15,6 +15,10 @@ Changelog
 - While redefining a signatory on an item, add possibility to select a
   `position_type` as label to use for the signature generated in POD templates.
   [gbastien]
+- Only call `MeetingItem._check_required_data` when item is about to be
+  presented into a meeting, this way previous transitions may be triggered by
+  configured process like Webservice call or when item sent from another cfg.
+  [gbastien]
 - Make the dashboard table header sticky so it is always viewable when
   scrolling, this is the case for every dashboards including
   `available/presented` items on the `meeting_view`.